refac: Introduce new base class for wholesale and energy result query base
