+++ source/OutgoingMessages.Infrastructure/Databricks/WholesaleResults/Queries/WholesaleAmountPerChargeQuery.cs
@@ -45,7 +45,7 @@ public class WholesaleAmountPerChargeQuery(
 
     public override string DataObjectName => "amounts_per_charge_v1";
 
-    public override string ActorColumnName => WholesaleResultColumnNames.EnergySupplierId;
+    protected override string ActorColumnName => WholesaleResultColumnNames.EnergySupplierId;
 
     public override Dictionary<string, (string DataType, bool IsNullable)> SchemaDefinition => new()
     {
