+++ source/OutgoingMessages.Infrastructure/Databricks/WholesaleResults/Queries/WholesaleMonthlyAmountPerChargeQuery.cs
@@ -44,7 +44,7 @@ public class WholesaleMonthlyAmountPerChargeQuery(
 
     public override string DataObjectName => "monthly_amounts_per_charge_v1";
 
-    public override string ActorColumnName => WholesaleResultColumnNames.EnergySupplierId;
+    protected override string ActorColumnName => WholesaleResultColumnNames.EnergySupplierId;
 
     public override Dictionary<string, (string DataType, bool IsNullable)> SchemaDefinition => new()
     {
