+++ source/OutgoingMessages.Infrastructure/Databricks/WholesaleResults/Queries/WholesaleTotalAmountQuery.cs
@@ -42,7 +42,7 @@ public class WholesaleTotalAmountQuery(
 
     public override string DataObjectName => "total_monthly_amounts_v1";
 
-    public override string ActorColumnName => WholesaleResultColumnNames.EnergySupplierId;
+    protected override string ActorColumnName => WholesaleResultColumnNames.EnergySupplierId;
 
     public override Dictionary<string, (string DataType, bool IsNullable)> SchemaDefinition => new()
     {
